+++ scripts/textAlign_Left.js
@@ -2,45 +2,70 @@
    textAlign_Left
 
    Description
-   This script changes the text alignment without moving the text position.
+   This script changes the text alignment without moving the position.
    Vertical text is also supported.
 
    Usage
-   Select the text objects, run this script from File > Scripts > Other Script...
+   Select point text objects, run this script from File > Scripts > Other Script...
 
    Notes
-   In rare cases, you may not be able to create it.
-   In that case, restart Illustrator and run this script again.
+   In rare cases, if you continue to use the script, it may not work.
+   In that case, restart Illustrator and try again.
 
    Requirements
    Illustrator CS or higher
 
    Version
-   1.0.0
+   1.0.1
 
    Homepage
    github.com/sky-chaser-high/adobe-illustrator-scripts
+
+   License
+   Released under the MIT license.
+   https://opensource.org/licenses/mit-license.php
    =============================================================================================================================================== */
 
 (function() {
-    if (app.documents.length) align(app.activeDocument.selection);
+    if (app.documents.length > 0 && app.activeDocument.selection.length > 0) main();
 })();
 
-function align(texts) {
+
+function main() {
+    var items = app.activeDocument.selection;
+    var texts = getTextFrames(items);
     for (var i = 0; i < texts.length; i++) {
-        if (texts[i].typename == 'TextFrame' && texts[i].kind == TextType.POINTTEXT) {
-            var left = texts[i].left;
-            var top = texts[i].top;
+        align(texts[i]);
+    }
+}
 
-            // work around a bug
-            texts[i].textRange.paragraphAttributes.justification = Justification.FULLJUSTIFYLASTLINELEFT;
 
-            texts[i].textRange.paragraphAttributes.justification = Justification.LEFT;
-            texts[i].left = left;
-            texts[i].top = top;
+function align(text) {
+    var position = text.position;
+
+    // work around a bug
+    // community.adobe.com/t5/illustrator-discussions/trouble-assigning-textframe-to-justification-left/m-p/4211277
+    var shrink = 80;
+    var expand = (1 / shrink) * 10000;
+
+    text.resize(shrink, shrink);
+    text.textRange.paragraphAttributes.justification = Justification.LEFT;
+    text.resize(expand, expand);
+
+    text.position = position;
+}
+
+
+function getTextFrames(items) {
+    var texts = [];
+    for (var i = 0; i < items.length; i++) {
+        var item = items[i];
+        if (item.typename == 'TextFrame' && item.kind == TextType.POINTTEXT) {
+            texts.push(item);
         }
-        else if (texts[i].typename == 'GroupItem') {
-            align(texts[i].pageItems);
+        if (item.typename == 'GroupItem') {
+            texts = texts.concat(getTextFrames(item.pageItems));
         }
     }
+    return texts;
 }